requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
dnspython==2.2.1
flake8==4.0.1
mccabe==0.6.1
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.1.1
PySocks==1.7.1
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rsa==4.8
six==1.16.0
Telethon==1.24.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
